+++ src/main/java/com/braintreegateway/TransactionLevelFeeReport.java
@@ -5,6 +5,7 @@ import java.io.InputStreamReader;
 import java.net.URL;
 import java.text.ParseException;
 import java.util.LinkedList;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.commons.csv.CSVFormat;
@@ -12,6 +13,7 @@ import org.apache.commons.csv.CSVRecord;
 
 public class TransactionLevelFeeReport {
     private List<TransactionLevelFeeReportRow> rows = new LinkedList<TransactionLevelFeeReportRow>();
+    private List<CSVRecord> csvRecords = new ArrayList<CSVRecord>();
     private Boolean valid;
 
     public TransactionLevelFeeReport(String urlValue) throws IOException, ParseException {
@@ -26,11 +28,21 @@ public class TransactionLevelFeeReport {
 
         for (CSVRecord record : CSVFormat.EXCEL.withFirstRecordAsHeader().parse(is)) {
             rows.add(new TransactionLevelFeeReportRow(record));
+            csvRecords.add(record);
         }
     }
 
+    /**
+     * Deprecated. Only supports the USA region. Please
+     * @see TransactionLevelFeeReport#getCSVRecords() which supports all regions.
+     */
+    @Deprecated
     public List<TransactionLevelFeeReportRow> getRows() {
         return rows;
+    }
+
+    public List<CSVRecord> getCSVRecords() {
+        return csvRecords;
     }
 
     public Boolean isValid() {