+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 * Add `getProcessorComments` to `Dispute`
 * Deprecate `getForwardedComments` from `Dispute`
+* Deprecate `getRows` from `TransactionLevelFeeReport` in favor of `getCSVRecords`
 
 ## 2.103.0
 * Add `getUniqueNumberIdentifier` to `CustomActionsPaymentMethodDetails`